+++ backend/src/api/post-api/dtos/block/write-block.dto.ts
@@ -1,14 +1,9 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { IsString, IsIn, IsLatitude, IsLongitude, IsOptional, ValidateNested } from 'class-validator';
-import { WriteBlockFileDto } from './write-block-files.dto';
+import { AttachFileDto } from '../file/attach-file.dto';
 
 export class WriteBlockDto {
-  @ApiProperty({ description: '블록을 수정하는 경우 UUID 식별자를 첨부합니다. 첨부하지 않으면 블록이 생성됩니다.' })
-  @IsOptional()
-  @IsString()
-  readonly uuid: string;
-
   @ApiProperty({ description: '블록을 설명할 수 있는 텍스트입니다.' })
   @IsString()
   readonly content: string;
@@ -31,13 +26,13 @@ export class WriteBlockDto {
   readonly longitude?: number;
 
   @ApiProperty({
-    type: WriteBlockFileDto,
+    type: AttachFileDto,
     isArray: true,
     required: false,
     description: '미디어 타입인 경우 필수로 파일 정보를 첨부합니다.',
   })
   @ValidateNested({ each: true })
-  @Type(() => WriteBlockFileDto)
+  @Type(() => AttachFileDto)
   @IsOptional()
-  files?: WriteBlockFileDto[];
+  files?: AttachFileDto[];
 }
