+++ backend/src/api/post-api/dtos/file/attach-file.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsOptional, IsUUID } from 'class-validator';
 
-export class WriteBlockFileDto {
+export class AttachFileDto {
   @ApiProperty({ description: '업로드한 파일의 식별자(UUID)를 첨부합니다.' })
   @IsUUID()
   readonly uuid: string;
